Implement UUID id for profile links
This is to prevent scrapers from indexing all
users. While the chance of collision is
astronomical, it is non-zero. There may be a
quick test check that I'm not aware of which
doesn't carry the performance cost of unique=True.

Note that in the event of collision, the user will
have to create a new account, as profile visits
will be sent to the profile for the
first uuid found.

I may just set unique=True for the time being,
though it is debatably unnecessary.