Fully implement GIF to PNG support
Really pleased with how this turned out.
This was another case of creating multiple files
when only one was needed. When I scrapped the
messy stuff and used just the files necessary,
it worked, and worked well. Goes to show that
when something isn't working, it can be good to
pivot towards something else that works and is
simpler.